custom_weather_alerts/requirements/production.txt
# Pro-tip: Try not to put anything here. Avoid dependencies in
#	production that aren't in development.
-r base.txt



# WSGI Handler
# ------------------------------------------------
gevent==1.2.2
gunicorn==19.7.1

# Static and Media Storage
# ------------------------------------------------
boto3==1.5.14
django-storages==1.6.5
Collectfast==0.6.0

# Email backends for Mailgun, Postmark, SendGrid and more
# -------------------------------------------------------
django-anymail==1.2

# Raven is the Sentry client
# --------------------------
raven==6.4.0


